t4demo/inventoryapp/serializer.py
from rest_framework import serializers


#Create your serializers here.
from .models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'
