t4demo/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-nose==1.4.7